Refactored the common components
